feat(ProjectListItemCollapsible): component composition step 1
