+++ src/components/ProjectListItemCollapsable/ProjectListItem.i18n/ru.json
